labs/week-24-job-scheduler/services/worker/requirements.txt
httpx==0.26.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation-httpx==0.43b0
prometheus-client==0.19.0
